INSP: move `Unused mut modifier` and `Unnecessary cast` inspections to `Rust/Lint` group
Also, move the corresponding classes into `inspections.lints` package

[skip ci]
